requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.2.0
Automat==22.10.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
boto3==1.26.49
botocore==1.29.49
bs4==0.0.1
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
cssselect==1.2.0
Django==4.1.4
django-cities==0.6.1
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-filter==22.1
django-model-utils==4.3.1
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
djangorestframework-simplejwt==5.2.2
dnspython==2.2.1
drf-yasg==1.21.4
drf-yasg-json-api==0.9.1
filelock==3.8.2
flower==1.2.0
humanize==4.4.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
itemadapter==0.7.0
itemloaders==1.0.6
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.2.4
lxml==4.9.2
MarkupSafe==2.1.1
ndg-httpsclient==0.5.1
packaging==22.0
parsel==1.7.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
Protego==0.2.1
psycopg2-binary==2.9.5
pusher==3.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.6
PyJWT==2.4.0
pymongo==4.3.3
PyNaCl==1.5.0
pyOpenSSL==22.1.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
queuelib==1.6.2
redis==4.4.1
redis-server==6.0.9
requests==2.28.1
requests-file==1.5.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
Scrapy==2.7.1
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
swapper==1.3.0
tldextract==3.4.0
tornado==6.2
tqdm==4.64.1
Twisted==22.10.0
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
vine==5.0.0
w3lib==2.1.1
wcwidth==0.2.5
whitenoise==6.3.0
zope.interface==5.5.2
